remove search image - upload images to project
